# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple Task Manager API",
        "title": "Task API",
        "contact": {
            "name": "shairayvonnecruz@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/tasks": {
            "get": {
                "description": "Get tasks",
                "tags": [
                    "V1"
                ],
                "summary": "Get all saved tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IndexResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "V1"
                ],
                "summary": "Create task in database",
                "parameters": [
                    {
                        "description": "Task Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaskRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "delete": {
                "description": "Delete task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "V1"
                ],
                "summary": "Delete task in database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update task",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "V1"
                ],
                "summary": "Update task in database",
                "parameters": [
                    {
                        "description": "Task Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TaskRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DeleteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.IndexResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.TaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.TaskResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Task"
                }
            }
        }
    }
}